daemons: rework the app label
Many kubernetes helper tools use this label to distinguish between
applications. Until now, all pods in the rucio-daemons chart where
assigned the same "app: rucio-daemons" label.

This is also problematic, and surprising that everything worked
correctly, because the deployment pod selector was the same for all
deployments, which is explicitly discouraged in the documentation:
https://kubernetes.io/docs/concepts/workloads/controllers/deployment/#selector

The new app label includes the daemon name: `rucio-daemons-conveyor-poller`
for example. The label `app-group` will now contain the old content of the
'app' label.