daemons: split common config and component config secrets
When I initially implemented it, seemed like a good idea to
have both the common and the per-daemon configuration in the
same secrets. This resulted in the duplication of the common
config in every daemon, but had the benefit of grouping the
data together when debugging was needed.

Now I changed my mind and I consider it's worth splitting the
two. In particular, to be able to use the common config in
non-daemon containers. For example, for fts renewal.