daemons: support per-component additionalSecrets
The commit allows to mount a secret only inside a specific
daemon. Until now, all secrets had to be mounted to all daemons.

This will allow, for example, to only hotfix patch a single
daemon. Another use case is to mount a specific secret,
proxy for example, only inside the daemons which need it.

Also rename additionalSecrets to secretMounts, but continue to
support the old naming too for the time being.