servers and daemons: rework sub-component extra secretMounts
In a very recent change, which shouldn't yet be used by anybody
except us, we introduced sub-component secretMounts. Those
were appended to whatever secretMounts where in the global config.
Unfortunately, this introduced a deadlock when the x509 proxy
was configured in the global config as a secret mount: the fts
cron job was trying to mount the x509 proxy before starting the
pod. And this is the job which actually generates the x509 proxy.

I decided to change the behavior of sub-component secretMounts.
Now they override the global secretMounts instead of extending it.
To add additional secretMounts to the global list, use
"extraSecretMounts".