server/daemons: Simplify restart command
The command is comprised of a pipeline to: (1) produce a list of
deployments, based on a label selector and (2) perform a rollout restart
for each of these deployments.  However, the label selector can be
passed directly to the rollout restart.  As such, there doesn’t appear a
valid reason for implementing this in such a way.

The entrypoint of the bitnami/kubectl container is ‘kubectl’, so it
feels preferable to pass the command-line arguments with ‘args’.